update deleteAllProducts function

+++ resources/views/cart.blade.php
@@ -7,27 +7,30 @@
             <h4 class="ps-2 ms-3 pt-3">Mi carrito de la compra</h4>
         </div>
 
+        <div class="ms-md-0 mt-3">
+            <a href="{{ route('home') }}">
+            Seguir comprando
+            </a>
+        </div>
+
         <!-- flash message -->
         <div class="ms-md-0 mt-3">
             @if (session()->has('message'))
             <div class="alert">
                 {{ session('message') }}
             </div>
-            <div>
-                <a href="{{ route('home') }}">
-                Clique aquí para seguir comprando
-                </a>
-            </div>
             @endif
         </div>
 
-        <div class="ms-md-0 mt-3">
-            <form action="{{route('deleteAllProducts')}}" method="post">
-                @method('delete')
-                @csrf
-                <button class="btn-deleteAllProducts bg-light" type="submit" onclick="return confirm('¿Estás seguro de que quieres eliminar todos los productos del carrito? ">Vaciar mi carrito</button>
-            </form>
-        </div>
+        @if ($products->count() > 0) 
+            <div class="ms-md-0 mt-3">
+                <form action="{{route('deleteAllProducts')}}" method="post">
+                    @method('delete')
+                    @csrf
+                    <button class="btn-deleteAllProducts bg-light" type="submit" onclick="return confirm('¿Estás seguro de que quieres eliminar todos los productos del carrito? ">Vaciar mi carrito</button>
+                </form>
+            </div>
+        @endif
 
     <div class="mt-md-3 ms-3 d-flex flex-wrap">
     @foreach ($products as $product)
